chtbot/pyproject.toml
[project]
name = "chtbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.9",
    "langchain-huggingface>=0.3.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pyngrok>=7.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.0",
    "langchain-pinecone>=0.2.11",
    "pinecone>=7.0.0",
    "faiss-cpu>=1.12.0",
    "sentence-transformers>=5.1.0",
    "nest-asyncio>=1.6.0",
]
